removing components that are not required
