--- blocks/form/components/file/file.css
@@ -1,1 +1,0 @@
-/* empty file.css */
