--- blocks/form/components/tnc/tnc.css
@@ -1,11 +1,0 @@
-.tnc__text-decoration {
-    color: #969696;
-    border: 1px solid #ccc;
-    border-radius: .25rem;
-    max-height: 160px;
-    margin-bottom: .25rem;
-    padding-left: 1rem;
-    padding-right: 1rem;
-    font-size: .875rem;
-    overflow: auto;
-}
