+++ blocks/form/submit.js
@@ -37,3 +37,85 @@ export function submitFailure(e, form) {
   form.setAttribute('data-submitting', 'false');
   form.querySelector('button[type="submit"]').disabled = false;
 }
+
+function generateUnique() {
+  return new Date().valueOf() + Math.random();
+}
+
+function getFieldValue(fe, payload) {
+  if (fe.type === 'radio') {
+    return fe.form.elements[fe.name].value;
+  } if (fe.type === 'checkbox') {
+    if (fe.checked) {
+      if (payload[fe.name]) {
+        return `${payload[fe.name]},${fe.value}`;
+      }
+      return fe.value;
+    }
+  } else if (fe.type !== 'file') {
+    return fe.value;
+  }
+  return null;
+}
+
+function constructPayload(form) {
+  const payload = { __id__: generateUnique() };
+  [...form.elements].forEach((fe) => {
+    if (fe.name && !fe.matches('button') && !fe.disabled && fe.tagName !== 'FIELDSET') {
+      const value = getFieldValue(fe, payload);
+      if (fe.closest('.repeat-wrapper')) {
+        payload[fe.name] = payload[fe.name] ? `${payload[fe.name]},${fe.value}` : value;
+      } else {
+        payload[fe.name] = value;
+      }
+    }
+  });
+  return { payload };
+}
+
+async function prepareRequest(form) {
+  const { payload } = constructPayload(form);
+  const headers = {
+    'Content-Type': 'application/json',
+  };
+  const body = { data: payload };
+  const url = form.dataset.submit || form.dataset.action;
+  return { headers, body, url };
+}
+
+export async function submitForm(form, captcha) {
+  try {
+    const { headers, body, url } = await prepareRequest(form, captcha);
+    let token = null;
+    if (captcha) {
+      token = await captcha.getToken();
+      body.data['g-recaptcha-response'] = token;
+    }
+    const response = await fetch(url, {
+      method: 'POST',
+      headers,
+      body: JSON.stringify(body),
+    });
+    if (response.ok) {
+      submitSuccess(
+        {
+          payload: {
+            body: {
+              thankYouMessage: form.dataset.thankYouMsg,
+              redirectUrl: form.dataset.redirectUrl,
+            },
+          },
+        },
+        form,
+      );
+    } else {
+      submitFailure({
+        payload: response,
+      }, form);
+    }
+  } catch (error) {
+    submitFailure({
+      payload: error,
+    }, form);
+  }
+}
